surprise/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy

# for cosine similarity
# scikit-learn

# for sobel similarity
# opencv-python